# docs/deliverables/external_design/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "OrderCreate": {
        "properties": {
          "coupon_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coupon Code"
          }
        },
        "title": "OrderCreate",
        "type": "object"
      },
      "OrderItemOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "product": {
            "$ref": "#/components/schemas/ProductOut"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "quantity",
          "price",
          "product"
        ],
        "title": "OrderItemOut",
        "type": "object"
      },
      "OrderOut": {
        "properties": {
          "coupon_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coupon Code"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "discount_amount": {
            "default": 0,
            "title": "Discount Amount",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/OrderItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "return_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Return Reason"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "total_price": {
            "title": "Total Price",
            "type": "number"
          },
          "user_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Email"
          }
        },
        "required": [
          "id",
          "total_price",
          "status",
          "created_at",
          "items"
        ],
        "title": "OrderOut",
        "type": "object"
      },
...
      "ProductImageOut": {
        "properties": {
          "display_order": {
            "title": "Display Order",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "title": "Image Url",
            "type": "string"
          }
        },
        "required": [
          "id",
          "image_url",
          "display_order"
        ],
        "title": "ProductImageOut",
        "type": "object"
      },
      "ProductOut": {
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "images": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProductImageOut"
            },
            "title": "Images",
            "type": "array"
          },
          "low_stock_threshold": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Low Stock Threshold"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "stock": {
            "title": "Stock",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "price",
          "stock",
          "id",
          "created_at"
        ],
        "title": "ProductOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "TechStore ECサイトのHTTP API。機械可読な正本はこのOpenAPI文書とする。",
    "title": "EC Site API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/payment/checkout": {
      "post": {
        "operationId": "create_checkout_session_payment_checkout_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Bearerトークンがない、無効、期限切れ、または退会済み"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Checkout Session"
      }
    },
    "/payment/complete": {
      "post": {
        "operationId": "complete_payment_payment_complete_post",
        "parameters": [
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Bearerトークンがない、無効、期限切れ、または退会済み"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Complete Payment"
      }
    },
...
  }
}